refactored double -> BigDecimal in User, Activity; refactored User attributes to be optional

+++ src/main/java/source/code/dto/request/activity/ActivityUpdateDto.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -18,7 +20,7 @@ public class ActivityUpdateDto {
     private String name;
 
     @Positive
-    private Double met;
+    private BigDecimal met;
 
     private Integer categoryId;
 }
